website/requirements.txt
tensorflow-macos==2.13.0
numpy==1.24.3
mne==1.4.2
scikit-learn==1.3.0
scipy==1.11.4
pytest==7.4.4